LCH-5434: Refactoring for events logging (#129)
